requirements.txt
numpy
pyserial
matplotlib
tqdm
pandas
lxml
scipy